% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas.R
\name{cal_aucell}
\alias{cal_aucell}
\title{Using AUCell to Calculate Pathway Activity Scores}
\usage{
cal_aucell(data, glist, n_cores = 4, seed.use = NULL, aucMaxRankPercent = 0.05)
}
\arguments{
\item{data}{normalized count matrix}

\item{glist}{gene set list}

\item{n_cores}{number of cores to used}

\item{seed.use}{seed for AUCell_buildRankings}

\item{aucMaxRankPercent}{percent of top genes to calculate AUC, common values may range from 0.01 to 0.2,
default is 0.05. It can also be set as 'auto', which will set aucMaxRank as the number of expressed genes
at least 99\% of cells had.}
}
\value{
A data frame of pathway activity scores, with cells in rows and pathways in columns.
}
\description{
Using AUCell to Calculate Pathway Activity Scores
}
