% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas.R
\name{cal_mean}
\alias{cal_mean}
\title{Using Mean as Pathway Activity Scores}
\usage{
cal_mean(data, glist, n_cores = 4, seed.use = NULL)
}
\arguments{
\item{data}{normalized count matrix}

\item{glist}{gene set list}

\item{n_cores}{number of cores to used}

\item{seed.use}{not used}
}
\value{
A data frame of pathway activity scores, with cells in rows and pathways in columns.
}
\description{
Using mean of expression of pathway genes as scores.
}
