% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas.R
\name{cal_pagoda2}
\alias{cal_pagoda2}
\title{Using Modified PAGODA2 to Calculate Pathway Activity Scores}
\usage{
cal_pagoda2(data, glist, n_cores = 4, seed.use = NULL)
}
\arguments{
\item{data}{normalized count matrix}

\item{glist}{gene set list}

\item{n_cores}{number of cores to used}

\item{seed.use}{not used}
}
\value{
A data frame of pathway activity scores, with cells in rows and pathways in columns.
}
\description{
Note::Default Pagoda2 method testPathwayOverdispersion was modified to testPathwayOverdispersion2, so
all input pathway scores (not just significant pathways) could be calculated.
}
