% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas.R
\name{cal_pas}
\alias{cal_pas}
\title{Functions using different pathway activity scoring methods.}
\usage{
cal_pas(
  data,
  glist,
  pas_method = "pagoda2",
  n_cores = 4,
  seed.use = NULL,
  aucMaxRankPercent = 0.05
)
}
\arguments{
\item{data}{normalized count matrix}

\item{glist}{gene set list}

\item{pas_method}{pathway activity scoring method used.
\itemize{
\item{pagoda2: }{default method, see \code{?cal_pagoda2}}
\item{default_pagoda2: }{see \code{?cal_default_pagoda2}}
\item{pagoda: }{see \code{?cal_pagoda}}
\item{aucell: }{see \code{?cal_aucell}}
\item{mean: }{see \code{?cal_mean}}
\item{ssgsea: }{see \code{?cal_ssgsea}}
\item{plage: }{see \code{?cal_plage}}
}}

\item{n_cores}{number of cores to used}

\item{seed.use}{only used for method "aucell"}

\item{aucMaxRankPercent}{percent of top genes to calculate AUC, only work for pas_method "aucell"}
}
\value{
A data frame of pathway activity scores, with cells in rows and pathways in columns.
}
\description{
Functions using different pathway activity scoring methods.
}
