% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{celltype_assign}
\alias{celltype_assign}
\title{Title}
\usage{
celltype_assign(
  markers,
  seu_obj,
  cell_score_method = "binary",
  cluster_score_method = "mean",
  cluster = "seurat_clusters"
)
}
\arguments{
\item{markers}{cell type marker used for identification. A data frame with two columns, "Markers" & "Celltype"}

\item{seu_obj}{seurat object used to identify cell type}

\item{cell_score_method}{one of "binary" or "raw"}

\item{cluster_score_method}{one of "freq" or "mean". If cell_score_method is "binary", only "mean" supported.
If cell_score_method is "raw", use "freq" as default, see detail.}

\item{cluster}{cell clusters to annotated, default is "seurat_clusters", which means seu_obj$seurat_clusters}
}
\value{
a data frame annotate each cluster to cell types
}
\description{
Title
}
\details{
For cell_score_method "binary", use data slot of seu_obj. Cells were scored as 1 if it expresses a marker (eg. counts > 0),
otherwise it was scored as 0. Then mean score across cells within a cluster and markers of a cell type were calcalated as
celltype score of a cluster. Clusters were assigned to cell type with the highest score.

For cell_score_method "raw", use scale.data slot of seu_obj, and cluster_score_method was set as "freq". Cells were scored
using the scaled expression of markers, and means of markers of each cell type were calculated as cell type scores.
Then, if method is \emph{freq}, annotate each cell to the cell type with highest scores, and annotate each cluster to
the cell type with most number of cells. If method is \emph{mean}, calculate the cluster scores as the means of cells of
the cluster, then annotate cluster to the cell type with highest scores.
}
