% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celltype_color}
\alias{celltype_color}
\title{Colors for Different Cell Type}
\format{
An object of class \code{character} of length 12.
}
\usage{
celltype_color
}
\description{
used for plot
}
\keyword{datasets}
