% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{clean_glist}
\alias{clean_glist}
\title{Clean Gene List}
\usage{
clean_glist(
  expressed.gene,
  used.gsets = NULL,
  glist = NULL,
  min.size = 5,
  suffix = TRUE
)
}
\arguments{
\item{expressed.gene}{genes expressed in your count matrix, can be \code{rownames(norm.data)}}

\item{used.gsets}{A data.frame represents gene set to be cleaned, can be
keggMetabDf, keggMetabDfSuper or reacMetabDf (see ?keggMetabDf for details).

You can also provide genes sets by yourself, make sure each row represents a genes, and columns "GeneSymbol"
(up-to-date HGNC symbols) and "Pathway" (Pathway names) are included.}

\item{glist}{provide gene sets as a list}

\item{min.size}{gene sets with genes less than min.size will be removed}

\item{suffix}{logical, whether to add gene set size as suffix to gene set name, default is TRUE}
}
\value{
a list object, each element represents a cleaned gene sets.
}
\description{
Filter gene list according to genes expressed (in your count matrix), you should filter genes in you matrix first!
}
\details{
Gene list could be provided as data.frame (to used.gsets) or list (to glist).
}
\examples{
genes <- c("HSD17B1", "PLA2G4A", "AACS", "MGLL", "MGST1")
# set min.size = 0 only for example
tmp <- clean_glist(genes, keggMetabDf, min.size = 0)
}
