% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ClusterPercent}
\alias{ClusterPercent}
\title{Feature Percent by Group}
\usage{
ClusterPercent(
  seu_obj,
  assay = NULL,
  features = NULL,
  group.by = NULL,
  threshold = 0
)
}
\arguments{
\item{seu_obj}{a seurat object}

\item{assay}{Which assays to use, default is \code{Seaurat::DefaultAssay(seu_obj)}}

\item{features}{features to calculate percent, default is all features in the assay}

\item{group.by}{Categories for grouping, default is not grouping, calculate the percent of whole dataset.}

\item{threshold}{threshold to detect expressed genes, default is 0}
}
\value{
a matrix of percents with row as features and column as groups
}
\description{
Calculate percent of each cells group which expressed specific features
}
