% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cor_heatmap}
\alias{cor_heatmap}
\title{Correlation Heatmap for Gene Expression or Reduction}
\usage{
cor_heatmap(
  seu_obj,
  reduction = NULL,
  metab.gene = NULL,
  dir.used = NULL,
  file.prefix = NULL,
  dims = 1:30,
  slot = "counts",
  vars = c("celltype", "patientID", "nFeature_metab", "nCount_metab"),
  split.by = "celltype",
  cor_method = "spearman",
  n = 500,
  seed.use = NULL,
  only.red = TRUE
)
}
\arguments{
\item{seu_obj}{seurat object}

\item{reduction}{Reduction name to be used to calculate cell-to-cell correlation, must can be found
in seu_obj. Default is NULL and metabolic gene expression will be used.}

\item{metab.gene}{Metabolic gene symbols to be used (if reduction is null). If not specified, it will
be set to genes in gaudeMetabDf.}

\item{dir.used}{directory used for *.pdf plot}

\item{file.prefix}{file prefix for *.pdf plot}

\item{dims}{Which dimensions to use to calculate correlation (if reduction is not null), default is 1:30.}

\item{slot}{which slot to use to fetch gene expression (if reduction is null), default is counts.}

\item{vars}{vars to set as heatmap row/column annotations, must included in seu_obj}

\item{split.by}{vars to split the plot, each will be plotted individually.}

\item{cor_method}{correlation method used, default is "spearman".}

\item{n}{number of cells to subsample for heatmap.}

\item{seed.use}{random seed used for subsample.}

\item{only.red}{only use red colors, default is TRUE}
}
\value{
*.pdf plot file
}
\description{
This function is used to plot cell-to-cell correlation heatmap based on (metabolic) gene expression
or reduction.
}
