% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_color}
\alias{dataset_color}
\title{Colors for Different datasets}
\format{
An object of class \code{character} of length 10.
}
\usage{
dataset_color
}
\description{
used for plot
}
\keyword{datasets}
