% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter Genes with High Drop-out Rate}
\usage{
filter_genes(seu_obj, percent_min = 5)
}
\arguments{
\item{seu_obj}{input seurat object}

\item{percent_min}{minimum proportion of cells expressing a gene, default is 5, means 5\% of cells.
Genes expressed in less than ncol(seu_obj) * percent_min / 100 cells will be removed.}
}
\value{
a seurat object with high drop-out genes removed.
}
\description{
Filter Genes with High Drop-out Rate
}
