% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaude_trans}
\alias{gaude_trans}
\title{Translate Gaude metabolic pathways to lower case (capitalized)}
\format{
An object of class \code{data.frame} with 96 rows and 2 columns.
}
\usage{
gaude_trans
}
\description{
Translate Gaude metabolic pathways to lower case (capitalized)
}
\keyword{datasets}
