% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaudeMetabDf}
\alias{gaudeMetabDf}
\title{Metabolic Genes and Pathways from Gaude et. al.}
\format{
A data frame with 1932 rows and 2 variables:
\describe{
\item{GeneSymbol}{Gene symbol}
\item{Pathway}{Pathway name}
}
}
\usage{
gaudeMetabDf
}
\description{
Manually curated metabolic pathways from Gaude et. al. (https://www.nature.com/articles/ncomms13041#Sec7)
}
\details{
Note: There \strong{are} duplicated genes in different pathway. Out-of-date gene symbols were updated
by \code{update_symbols()}
}
\keyword{datasets}
