% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Split violin plot}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data to displayed}

\item{stat}{stat}

\item{position}{positon}

\item{...}{...}

\item{draw_quantiles}{draw_quantiles}

\item{trim}{trim}

\item{scale}{scale}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}
}
\value{
split violin plot
}
\description{
see ?geom_violin for detail
}
