% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpdatedSymbols.R
\name{getUpdatedSymbols_searat}
\alias{getUpdatedSymbols_searat}
\title{Updated Gene Symbols for Seurat Object}
\usage{
getUpdatedSymbols_searat(seu_obj)
}
\arguments{
\item{seu_obj}{Input seurat object}
}
\value{
seurat object with gene symbols updated.
}
\description{
Seurat do not allowed renaming features, so get counts data out, rename it and reconstructed seurat object.
}
