% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HallmarksDf}
\alias{HallmarksDf}
\alias{cancerHallmarksDf}
\title{Cancer Hallmark Genes and Pathways from MSigDB.}
\format{
\describe{
\item{GeneSymbol}{Gene symbol}
\item{Pathway}{Pathway name}
}

An object of class \code{data.frame} with 1147 rows and 2 columns.
}
\usage{
HallmarksDf

cancerHallmarksDf
}
\description{
Cancer Hallmark Genes and Pathways from MSigDB.
\code{HallmarksDf} contains all MSigDB hallmark gene sets
\code{cancerHallmarksDf} contains 6 cancer hallmarks and mTORC1 signaling pathway
}
\details{
Note: There \strong{are} duplicated genes in different pathway. Out-of-date gene symbols were updated
by \code{update_symbols()}
}
\keyword{datasets}
