% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgnc}
\alias{hgnc}
\title{Approved HGNC Gene Symbols}
\format{
A data frame with 42571 rows and 9 variables:
\describe{
\item{Approved.symbol}{The official gene symbol that has been approved by the HGNC and is publicly available.
Symbols are approved based on specific HGNC nomenclature guidelines. }
\item{Previous.symbols}{Symbols previously approved by the HGNC for this gene.
This field can contain multiple values as a comma delimited list.}
}
}
\source{
\url{https://www.genenames.org/download/custom/}
}
\usage{
hgnc
}
\description{
A dataset containing the HGNC gene symbol information, which is download from
"https://www.genenames.org/download/custom/" with "Select status" only include "Approved"
in Jun 1, 2021, saved as "data-raw/hgnc_20210601.txt"
}
\keyword{datasets}
