% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ident_malig}
\alias{ident_malig}
\title{Distinguish Malignant and Non-malignant Cells Iteratively}
\usage{
ident_malig(
  seu_obj,
  init_list,
  out_pdf = "ident_malignant_iteration.pdf",
  max_iter = 20,
  len_feature_list = 50,
  max.cells.per.ident = 2000,
  error_rate = 0.001
)
}
\arguments{
\item{seu_obj}{seurat object contain epithelial cells}

\item{init_list}{Initial list of malignant and non_malignant features. A list with two elements
named as "malignant" and "non_malignant", each contain 50 features.}

\item{out_pdf}{pdf file for clustering plot of each iteration.}

\item{max_iter}{maximum iterations, default is 20}

\item{len_feature_list}{length of feature list to calculate score. Default if 50}

\item{max.cells.per.ident}{max cells per ident for \code{Seurat::FindMarkers}, default is 2000}

\item{error_rate}{is the portion of misclassified cell compared to last iteration if less than
error_rate, then it is regarded as convergent.}
}
\value{
a data.frame of clustering result for each iteration
}
\description{
Distinguish Malignant and Non-malignant Cells Iteratively
}
\details{
First, assign an initial malignant score and non-malignant score for each epithelial cell in
\code{seu_obj} according to init_list via \code{Seurat::AddModuleScore}. Then, k-means clustering algorithm
will be applied on the two scores to define malignant cells and non-malignant cells. Next,
Differentially expressed genes between this putative malignant and non-malignant was identified
as new features of malignant and non-malignant (50 features each). Finally, repeat the scoring
and clustering iteratively until the portion of misclassified cells less than \code{error_rate}.
}
