% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immuneDf}
\alias{immuneDf}
\title{Immune Genes}
\format{
\describe{
\item{GeneSymbol}{Gene symbol}
\item{Pathway}{Pathway name}
}
}
\usage{
immuneDf
}
\description{
Immune Genes
}
\details{
Note: There \strong{are} duplicated genes in different pathway. Out-of-date gene symbols were updated
by \code{update_symbols()}
}
\keyword{datasets}
