% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jaccard_dist}
\alias{jaccard_dist}
\title{Jaccard Distance of Two Gene Sets}
\usage{
jaccard_dist(x, y)
}
\arguments{
\item{x}{gene sets 1 (gs1)}

\item{y}{gene sets 2 (gs2)}
}
\value{
distance in [0, 1]
}
\description{
Jaccard Distance of Two Gene Sets
}
\examples{
jaccard_dist(c("PTEN", "TP53"), c("PTEN", "DUSP2"))
}
