% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keggMetabDf}
\alias{keggMetabDf}
\title{Metabolic Genes and Pathways from KEGG}
\format{
A data frame with 2958 rows and 10 variables:
\describe{
\item{CateID}{KEGG category ID}
\item{Category}{KEGG category name}
\item{PathID}{KEGG pathway ID}
\item{Pathway}{KEGG pathway name}
\item{GeneID}{NCBI gene ID}
\item{GeneSymbol}{HGNC gene symbol}
\item{GeneDescription}{Gene description}
\item{ProteinID}{KEGG protein ID}
\item{ProteinName}{}
\item{ProteinDescription}{}
...
}
}
\usage{
keggMetabDf
}
\description{
A dataset containing metabolic genes information from KEGG metabolic pathways, generated from
"data-raw/hsa00001.json" files downloaded from KEGG website, which include all kegg pathway info of homo sapiens.
}
\details{
Note: There \strong{are} duplicated genes in different pathway.
}
\keyword{datasets}
