% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keggMetabDfSuper}
\alias{keggMetabDfSuper}
\title{Metabolic Genes and Super Pathways from KEGG}
\format{
A data frame with 2958 rows and 10 variables:
\describe{
\item{Category}{KEGG categories}
\item{Pathway}{KEGG super pathway name}
\item{GeneSymbol}{HGNC gene symbol}
}
}
\usage{
keggMetabDfSuper
}
\description{
A dataset containing metabolic genes information from KEGG metabolic pathways, generated from
"data-raw/hsa00001.json" files downloaded from KEGG website, which include all kegg pathway info of homo sapiens.

Note: use categories as super pathways
}
\details{
Note: There \strong{are} duplicated genes in different super pathway.
}
\keyword{datasets}
