% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myeloidSig}
\alias{myeloidSig}
\title{Signature Genes of Myeloid}
\format{
\describe{
\item{GeneSymbol}{Gene symbol}
\item{Pathway}{Pathway name}
}
}
\usage{
myeloidSig
}
\description{
This data frame contains signatures of myeloid subtype identified by Cheng et al.
}
\details{
Note: There \strong{are} duplicated genes in different pathway. Out-of-date gene symbols were updated
by \code{update_symbols()}
}
\references{
Cheng, L. et al. A pan-cancer single-cell transcriptional atlas of tumor infiltrating myeloid cells. Cell (2021)
doi:10.1016/j.cell.2021.01.010
}
\keyword{datasets}
