% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pathway_correlation}
\alias{pathway_correlation}
\title{Pathway Correlation and P Value}
\usage{
pathway_correlation(
  seu_obj = NULL,
  assay = "AUCell",
  features = NULL,
  df = NULL,
  cor_method = "spearman",
  glist = NULL
)
}
\arguments{
\item{seu_obj}{a Seuat object with an assay stored pathway activity score (PAS)}

\item{assay}{assay name of PAS, default is "AUCell"}

\item{features}{pathways to calculate pair-wise correlation}

\item{df}{provide a data frame of pathway scores instead of a seurat object}

\item{cor_method}{one of "spearman" or "pearson", default is "spearman"}

\item{glist}{gene set list to correspond to PAS, used to calculate pathway distance}
}
\value{
A list containing the following components:
r    the matrix of correlations
n    the matrix of number of observations used in analyzing each pair f variables
P    the asymptotic P-values
dist the matrix of jaccard distance
radj correlations adjusted by dist
padj adjusted P-values
}
\description{
Pathway Correlation and P Value
}
