% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pathway_gene_heatmap}
\alias{pathway_gene_heatmap}
\title{DoHeatmap of Pathwa Genes}
\usage{
pathway_gene_heatmap(
  seu_obj,
  pathway_genes,
  group.by = "celltype",
  cells = NULL,
  colors = NULL
)
}
\arguments{
\item{seu_obj}{seurat object, must include RNA assay}

\item{pathway_genes}{gene names of a pathway}

\item{group.by}{variable names to group the heatmap by, default is "celltype"}

\item{cells}{cells to used}

\item{colors}{a vector of colors for different groups}
}
\value{
a heatmap
}
\description{
Re-scale seurat object using specified pathway_genes
}
