% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{pathway_seurat}
\alias{pathway_seurat}
\title{Title}
\usage{
pathway_seurat(pas, meta_data, file_prefix = NULL, out_dir = "./", ...)
}
\arguments{
\item{pas}{PAS matrix or data.frame, cells in rows and pathways in columns}

\item{meta_data}{cell meta info}

\item{file_prefix}{file prefix for .pdf plot file}

\item{out_dir}{output directory, default is "./"}

\item{...}{other parameters}
}
\value{
cell type markers
}
\description{
Title
}
