% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{pathway_seurat_clustering}
\alias{pathway_seurat_clustering}
\title{Seurat Clustering for Pathway Activity Score (PAS)}
\usage{
pathway_seurat_clustering(
  pas,
  meta_data,
  meta_var_used = c("nCount_RNA", "nFeature_RNA", "percent.mt", "celltype",
    "sub_celltype", "patientID", "S.Score", "G2M.Score", "Phase"),
  n_pcs = 5,
  resolution = 0.1,
  seed.use = 2021
)
}
\arguments{
\item{pas}{PAS matrix or data.frame, cells in rows and pathways in columns}

\item{meta_data}{cell meta info}

\item{meta_var_used}{fields assumed to exist in meta_data, including:
\itemize{
\item{nCount_RNA: }{total UMI counts of the cell}
\item{nFeature_RNA: }{number of genes expressed in the cell}
\item{percent.mt: }{percentage of counts from mitochondria genes in total UMI counts}
\item{celltype: }{identified main cell types}
\item{sub_celltype: }{identified cell subtypes, optional}
\item{patientID: }{patient ID}
\item{S.Score, G2M.Score, Phase: }{results of Seurat::CellCycleScoring()}
}}

\item{n_pcs}{number of PC used for clustering, default is 5}

\item{resolution}{resolution for Seurat::FindClusters(), default is 0.1}

\item{seed.use}{random seed used, default is 2021}
}
\value{
a clustered Seurat object with an assay named "PAS"
}
\description{
Seurat Clustering for Pathway Activity Score (PAS)
}
