% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{pathway_seurat_markers}
\alias{pathway_seurat_markers}
\title{Title}
\usage{
pathway_seurat_markers(sc, group.by = "celltype", only.pos = TRUE)
}
\arguments{
\item{sc}{a Seurat object constructed with PAS data}

\item{group.by}{variables to group cell for comparing, default is "celltype"}

\item{only.pos}{if only find positive markers, default is TRUE}
}
\value{
cell markers of different cell types
}
\description{
Title
}
