% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pathway_seurat_plot}
\alias{pathway_seurat_plot}
\title{Title}
\usage{
pathway_seurat_plot(
  sc,
  file_prefix = NULL,
  out_dir = "./",
  reduction_type = c("tsne", "umap"),
  group_type = c("celltype", "seurat_clusters", "patientID", "Phase"),
  qc_field = c("nFeature_RNA", "nCount_RNA", "percent.mt")
)
}
\arguments{
\item{sc}{a Seurat object constructed with PAS data}

\item{file_prefix}{file prefix of .pdf plot file}

\item{out_dir}{directory for output}

\item{reduction_type}{reduction type used}

\item{group_type}{variable to group cells}

\item{qc_field}{QC field to plot using Seurat::FeaturePlot()}
}
\description{
Title
}
