% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCNV}
\alias{plotCNV}
\title{CNV Heatmap}
\usage{
plotCNV(
  infercnv,
  meta_info,
  split.by,
  group.by = NULL,
  max.cells.per.ident = 100,
  cluster_rows = FALSE,
  use_raster = FALSE,
  name = "CNV scores"
)
}
\arguments{
\item{infercnv}{infercnv object}

\item{meta_info}{cell meta info}

\item{split.by}{column names in meta_info to split cells, prior to group.by}

\item{group.by}{column names in meta_info to group cells}

\item{max.cells.per.ident}{maximum cells per identity to plot}

\item{cluster_rows}{whether to cluster rows for heatmap}

\item{use_raster}{whether use raster}

\item{name}{plot name}
}
\value{
a ComplexHeatmap plot
}
\description{
CNV Heatmap
}
