% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPath_x_libsize}
\alias{plotPath_x_libsize}
\title{Scatter plot of pathway activity score - library size}
\usage{
plotPath_x_libsize(
  pas,
  pas.tools = NULL,
  meta.data,
  libsize_type = "nCount_RNA",
  out.dir
)
}
\arguments{
\item{pas}{a list containing different type of pathway activity score}

\item{pas.tools}{names of pas tools to plot, if NULL then set as names(pas)}

\item{meta.data}{a data.frame containing metadata of cells}

\item{libsize_type}{variable names of library size, one of "nCount_RNA" of "nFeature_RNA"}

\item{out.dir}{output directory for plot file}
}
\description{
Scatter plot of pathway activity score - library size
}
