% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotPath_x_Path}
\alias{plotPath_x_Path}
\title{Pathway-Pathway Plot from Different PAS Methods}
\usage{
plotPath_x_Path(pas, out.dir = NULL, group.by = NULL)
}
\arguments{
\item{pas}{a list containing different type of pathway activity score}

\item{out.dir}{output directory for plot file}

\item{group.by}{variable used to group data points}
}
\description{
Pathway-Pathway Plot from Different PAS Methods
}
