% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reacMetabDf}
\alias{reacMetabDf}
\title{Metabolic Genes and Pathways from Reactome}
\format{
A data frame with 1916 rows and 2 variables:
\describe{
\item{GeneSymbol}{Gene symbol}
\item{Pathway}{belong to which one of 7 super pathways}
}
}
\usage{
reacMetabDf
}
\description{
Seven metabolic super-pathways based on the latest Reactome annotations (Fabregat et al., 2016) were
curated by \href{https://www.cell.com/cell-reports/fulltext/S2211-1247(18)30438-8}{Peng et al., 2018}.
Included were amino acid metabolism (348 genes), carbohydrate metabolism (286 genes),
integration of energy (110 genes), lipid metabolism (766 genes), nucleotide metabolism (90 genes),
tricarboxylic acid cycle (TCA cycle, 148 genes) and vitamin & cofactor metabolism (168 genes).
Generated using Table S2 (save as "data-raw/Reactome_metab_TableS2.xlsx") from this article.
}
\details{
Note: There \strong{are} duplicated genes in different pathway. Out-of-date gene symbols were updated
by \code{getUpdatedSymbols()}
}
\keyword{datasets}
