% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{reorder_cluster_res}
\alias{reorder_cluster_res}
\title{Reorder Cluster Results of Seurat}
\usage{
reorder_cluster_res(seu_obj, resolution, assay_name = "RNA")
}
\arguments{
\item{seu_obj}{input seurat object}

\item{resolution}{resolutions of cluster results. The cluster results of Seurat were names as "{assay_name}_snn_res.{resolution}"}

\item{assay_name}{assay_name of cluster results}
}
\value{
a reorder seurat object
}
\description{
The clusters results identified by Seurat::FindClusters were a factor with levels from 1 to n.
However, when singletons (which is cluster consist of only one cell) produced during cluster
identifications, the levels of the factor will sort as character (eg: just like "0", "1", "10", "2", "3", ...).
This function was used to reorder cluster results as integer (eg: just like "0", "1", "2", "3", ..., "9", "10", "11")
}
