% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_cell_pairs}
\alias{sample_cell_pairs}
\title{Sample Cell Pairs for Comparison}
\usage{
sample_cell_pairs(
  seu_obj,
  group.by = "patientID",
  split.by = "celltype",
  sample_type = c("inter", "intra"),
  n = 1000,
  seed.use = NULL
)
}
\arguments{
\item{seu_obj}{seurat object}

\item{group.by}{variables to group cells, cell from inter-group and intra-group will be compared}

\item{split.by}{variables to split cells, cells will be split first and then grouped individually}

\item{sample_type}{weather to sample inter-group cells or intra-group cells, default is both}

\item{n}{number of cell pairs for each split}

\item{seed.use}{seed used for subsample}
}
\value{
a data.frame of select cell pairs
}
\description{
Down sample cell pairs from inter-group and intra-group for comparison
}
