% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{seurat_pipe}
\alias{seurat_pipe}
\title{Default Seurat Clustering Pipeline}
\usage{
seurat_pipe(
  seu,
  features = NULL,
  vars.to.regress = NULL,
  n_pcs = 30,
  useHarmony = FALSE,
  harmony.group.var = "samID",
  max.iter.hamony = 30,
  reduction = "pca",
  clustering = TRUE,
  resolution = c(0.2, 0.5, 0.8, 1.2),
  seed.use = 2021,
  useTSNE = TRUE,
  verbose = TRUE,
  fast_tsne_path = NULL
)
}
\arguments{
\item{seu}{a Seurat object}

\item{features}{Vector of features names to scale/center. Default is variable features.}

\item{vars.to.regress}{variables to regress out. For example, nUMI, or precent.mt}

\item{n_pcs}{number of PCs used. Default is 30}

\item{useHarmony}{whether to use harmony to integrate cells from different source. Default is FALSE}

\item{harmony.group.var}{var to integrate, must be specified if useHarmony is TRUE.}

\item{max.iter.hamony}{maximum iterations for harmony}

\item{reduction}{reduction used for dimensional reduction and clustering. Default is "pca". If
useHarmony is TRUE, it will be set as "harmony"}

\item{clustering}{do clustering or not. Default is TRUE.}

\item{resolution}{resolution parameter for Seurat::FindClusters}

\item{seed.use}{random seed used. Default is 2021}

\item{useTSNE}{whether to do TSNE dimensional reduction. Default is TRUE}

\item{verbose}{control verbosity. Default is TRUE}

\item{fast_tsne_path}{for windows, specify the path of fastTSNE.exe}
}
\value{
A Seurat object with clustering done.
}
\description{
Packaging seurat default clustering pipeline in a single function.
}
