% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcellSig}
\alias{tcellSig}
\title{Signature Genes of T Cell Subtype}
\format{
\describe{
\item{GeneSymbol}{Gene symbol}
\item{Pathway}{Pathway name}
}
}
\usage{
tcellSig
}
\description{
This data frame contains top 50 signature genes for each T cell subtype identified by Zheng et al. They identified 17 CD8+ and
24 CD4+ metaclusters and corresponding signatures.
}
\details{
Note: There \strong{are} duplicated genes in different pathway. Out-of-date gene symbols were updated
by \code{update_symbols()}
}
\references{
Zheng, L. et al. Pan-cancer single-cell landscape of tumor-infiltrating T cells. Science (2021) doi:10.1126/science.abe6474.
}
\keyword{datasets}
