% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TFgenes}
\alias{TFgenes}
\title{Human Transcriptional Factor Coding Genes}
\format{
An object of class \code{character} of length 1839.
}
\usage{
TFgenes
}
\description{
1839 genes, with symbols updated use upadted_symbols()
}
\keyword{datasets}
