% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TvN_color}
\alias{TvN_color}
\title{Colors for Tumor and Normal}
\format{
An object of class \code{character} of length 2.
}
\usage{
TvN_color
}
\description{
used for plot
}
\keyword{datasets}
