% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpdatedSymbols.R
\name{update_symbols_searat}
\alias{update_symbols_searat}
\title{Updated Gene Symbols for Seurat Object}
\usage{
update_symbols_searat(seu_obj)
}
\arguments{
\item{seu_obj}{Input seurat object}
}
\value{
seurat object with gene symbols updated.
}
\description{
Seurat do not allowed renaming features, so get counts data out, rename it and reconstructed seurat object.
}
