% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpdatedSymbols.R
\name{update_symbols_seurat}
\alias{update_symbols_seurat}
\title{Updated Gene Symbols for Seurat Object}
\usage{
update_symbols_seurat(seu_obj, out_dir = "res/")
}
\arguments{
\item{seu_obj}{Input seurat object}

\item{out_dir}{output dir for gene symbol updated table}
}
\value{
seurat object with gene symbols updated.
}
\description{
Seurat do not allowed renaming features, so get counts data out, rename it and reconstructed seurat object.
}
