% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas.R
\name{cal_AUCell}
\alias{cal_AUCell}
\title{Using AUCell to Calculate Pathway Activity Scores}
\usage{
cal_AUCell(data, glist, n_cores)
}
\arguments{
\item{data}{normalized count matrix}

\item{glist}{gene set list}

\item{n_cores}{number of cores to used}
}
\value{
A data frame of pathway activity scores, with cells in rows and pathways in columns.
}
\description{
Using AUCell to Calculate Pathway Activity Scores
}
