% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_glist}
\alias{clean_glist}
\title{Clean Gene List}
\usage{
clean_glist(expressed.gene, used.gsets, min.size = 5, suffix = TRUE)
}
\arguments{
\item{expressed.gene}{genes expressed in your count matrix, can be \code{rownames(norm.data)}}

\item{used.gsets}{A data.frame, gene sets to be cleaned, can be keggMetabDf or reacMetabDf. You can also provide genes sets
by yourself, make sure each row represents a genes, and columns "GeneSymbol" (up-to-date HGNC symbols) and
"Pathway" (Pathway names) are included.}

\item{min.size}{gene sets with genes less than min.size will be removed}

\item{suffix}{logical, whether to add gene set size as suffix to gene set name, default is TRUE}
}
\value{
a list object, each element represents a cleaned gene sets.
}
\description{
Filter gene list according to genes expressed (in your count matrix), you should run filter_genes() first!
}
