% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgnc}
\alias{hgnc}
\title{Approved HGNC Gene Symbols}
\format{
A data frame with 42571 rows and 9 variables:
\describe{
\item{HGNC.ID}{A unique ID provided by the HGNC}
\item{Approved.symbol}{The official gene symbol that has been approved by the HGNC and is publicly available.
Symbols are approved based on specific HGNC nomenclature guidelines. }
\item{Approved.name}{The official gene name that has been approved by the HGNC and is publicly available.
Names are approved based on specific HGNC nomenclature guidelines.}
\item{Status}{Indicates whether the gene is classified as: Approved, Entry withdrawn or Symbol withdrawn.
only Approved ones are included in this dataset.}
\item{Previous.symbols}{Symbols previously approved by the HGNC for this gene.
This field can contain multiple values as a comma delimited list.}
\item{Alias.symbols}{Other symbols used to refer to this gene.
This field can contain multiple values as a comma delimited list.}
\item{Chromosome}{Indicates the location of the gene or region on the chromosome.}
\item{Accession.numbers}{Accession numbers for each entry selected by the HGNC.
This field can contain multiple values as a comma delimited list.}
\item{RefSeq.IDs}{The Reference Sequence (RefSeq) identifier for that entry, provided by the NCBI.}
...
}
}
\source{
\url{https://www.genenames.org/download/custom/}
}
\usage{
hgnc
}
\description{
A dataset containing the HGNC gene symbol information, which is download from
"https://www.genenames.org/download/custom/" with "Select status" only include "Approved"
in Jun 1, 2021, saved as "data-raw/hgnc_20210601.txt"
}
\keyword{datasets}
