% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotPath_x_libsize}
\alias{plotPath_x_libsize}
\title{Title}
\usage{
plotPath_x_libsize(pas, meta.data, out.dir)
}
\arguments{
\item{pas}{a list containing different type of pathway activity score}

\item{meta.data}{a data.frame containing metadata of cells}

\item{out.dir}{output directory for plot file}
}
\description{
Title
}
